+++ src/preset/node/providers/express-factory.ts
@@ -1,0 +1,9 @@
+import express from 'express';
+
+/**
+ * Провайдер фабрики http-серверов.
+ * @return Фабрика.
+ */
+export function provideExpressFactory() {
+  return express;
+}
