+++ src/preset/node/providers/express-request-middleware.ts
@@ -1,0 +1,10 @@
+import type express from 'express';
+import { Handlers } from '@sentry/node';
+
+/**
+ * Провайдер промежуточного слоя учета входящих http-запросов.
+ * @return Промежуточный слой.
+ */
+export function provideExpressRequestMiddleware(): express.Handler {
+  return Handlers.requestHandler();
+}
